Added UI for multinode configuration on Accounts location
