+++ sclient/src/main/scala/synereo/client/css/UserProfileViewCSS.scala
@@ -43,6 +43,16 @@ object UserProfileViewCSS {
       color.white
     )
 
+    val editSaveButton = style(
+      ToStyleAV(marginLeft(1.%%)),
+      marginTop(1.px)
+    )
+
+    val inputText = style(
+      width(200.px),
+      display.inline
+    )
+
   }
 
 }
